Trying to resolve Issue #3 - converting references to t_mean to t_diff.  We've talked about this a bit.
Added a block of commented code that I used to change the annual values in `climate_ts`.  `load_climate` changed to explicitly pull out the diffed annual values, had to change the ranges in `resampling_climate` as well.

Changed some variable names in `climate_ts` to reflect the new standard of using an underscore instead of a period in variable names.  This does not yet happen throughout.