Began removing some of the *pathogen* content, or at least it's emphasis in the purpose of the paper.
Began changing tense in the methods to present where appropriate based on the advice of Reviewer 3.

Better explanation of the terminology later used in figure captions, based on the advice of Reviewer 3.

Explained why we used the normal we did.

Added in the Hellinger equation and rephrased the explanation of how it was used.

Added some packages into the **Analysis** section.

Added references to Rhemtulla et al 2007 in the Discussion.